+++ packs/efeitos-cardigan/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-001316
+MANIFEST-001318
