--- packs/efeitos-cardigan/LOG
@@ -1,3 +1,0 @@
-2026/07/01-22:16:16.791 3008 Recovering log #1314
-2026/07/01-22:16:16.843 3008 Delete type=0 #1314
-2026/07/01-22:16:16.843 3008 Delete type=3 #1312
