+++ packs/efeitos-cardigan/LOG.old
@@ -1,8 +1,3 @@
-2026/07/01-22:10:33.360 3008 Recovering log #1310
-2026/07/01-22:10:33.434 3008 Delete type=0 #1310
-2026/07/01-22:10:33.434 3008 Delete type=3 #1308
-2026/07/01-22:15:45.352 3764 Level-0 table #1315: started
-2026/07/01-22:15:45.352 3764 Level-0 table #1315: 0 bytes OK
-2026/07/01-22:15:45.394 3764 Delete type=0 #1313
-2026/07/01-22:15:45.466 3764 Manual compaction at level-0 from '!folders!FszDz3GuPvOHfi8w' @ 72057594037927935 : 1 .. '!items!xY2yZ3zA4bB5cC6D' @ 0 : 0; will stop at (end)
-2026/07/01-22:15:45.466 3764 Manual compaction at level-1 from '!folders!FszDz3GuPvOHfi8w' @ 72057594037927935 : 1 .. '!items!xY2yZ3zA4bB5cC6D' @ 0 : 0; will stop at (end)
+2026/07/01-22:16:16.791 3008 Recovering log #1314
+2026/07/01-22:16:16.843 3008 Delete type=0 #1314
+2026/07/01-22:16:16.843 3008 Delete type=3 #1312
