+++ packs/equipamentos-cardigan/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-001327
+MANIFEST-001329
