--- packs/equipamentos-cardigan/LOG
@@ -1,3 +1,0 @@
-2026/07/01-22:16:16.958 39a0 Recovering log #1325
-2026/07/01-22:16:17.020 39a0 Delete type=0 #1325
-2026/07/01-22:16:17.020 39a0 Delete type=3 #1323
