+++ packs/equipamentos-cardigan/LOG.old
@@ -1,8 +1,3 @@
-2026/07/01-22:10:33.606 3798 Recovering log #1321
-2026/07/01-22:10:33.661 3798 Delete type=0 #1321
-2026/07/01-22:10:33.662 3798 Delete type=3 #1319
-2026/07/01-22:15:45.418 3764 Level-0 table #1326: started
-2026/07/01-22:15:45.418 3764 Level-0 table #1326: 0 bytes OK
-2026/07/01-22:15:45.444 3764 Delete type=0 #1324
-2026/07/01-22:15:45.466 3764 Manual compaction at level-0 from '!folders!A1cE2sS3oR4iO5sF' @ 72057594037927935 : 1 .. '!items!Z4qG7jL1mW5sN9vP' @ 0 : 0; will stop at (end)
-2026/07/01-22:15:45.466 3764 Manual compaction at level-1 from '!folders!A1cE2sS3oR4iO5sF' @ 72057594037927935 : 1 .. '!items!Z4qG7jL1mW5sN9vP' @ 0 : 0; will stop at (end)
+2026/07/01-22:16:16.958 39a0 Recovering log #1325
+2026/07/01-22:16:17.020 39a0 Delete type=0 #1325
+2026/07/01-22:16:17.020 39a0 Delete type=3 #1323
