+++ packs/racas-cardigan/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-002275
+MANIFEST-002277
