--- packs/racas-cardigan/LOG
@@ -1,3 +1,0 @@
-2026/07/01-22:16:16.899 3798 Recovering log #2273
-2026/07/01-22:16:16.953 3798 Delete type=0 #2273
-2026/07/01-22:16:16.953 3798 Delete type=3 #2271
