+++ packs/racas-cardigan/LOG.old
@@ -1,8 +1,3 @@
-2026/07/01-22:10:33.534 39a0 Recovering log #2269
-2026/07/01-22:10:33.602 39a0 Delete type=0 #2269
-2026/07/01-22:10:33.602 39a0 Delete type=3 #2267
-2026/07/01-22:15:45.444 3764 Level-0 table #2274: started
-2026/07/01-22:15:45.444 3764 Level-0 table #2274: 0 bytes OK
-2026/07/01-22:15:45.465 3764 Delete type=0 #2272
-2026/07/01-22:15:45.466 3764 Manual compaction at level-0 from '!items!D6eF9gG2hI5iJ8kL' @ 72057594037927935 : 1 .. '!items!V9wX2yY5zZ8aA1bB' @ 0 : 0; will stop at (end)
-2026/07/01-22:15:45.466 3764 Manual compaction at level-1 from '!items!D6eF9gG2hI5iJ8kL' @ 72057594037927935 : 1 .. '!items!V9wX2yY5zZ8aA1bB' @ 0 : 0; will stop at (end)
+2026/07/01-22:16:16.899 3798 Recovering log #2273
+2026/07/01-22:16:16.953 3798 Delete type=0 #2273
+2026/07/01-22:16:16.953 3798 Delete type=3 #2271
