+++ packs/skills-cardigan/CURRENT
@@ -1,1 +1,1 @@
-MANIFEST-001344
+MANIFEST-001346
