--- packs/skills-cardigan/LOG
@@ -1,3 +1,0 @@
-2026/07/01-22:16:16.848 40cc Recovering log #1342
-2026/07/01-22:16:16.894 40cc Delete type=0 #1342
-2026/07/01-22:16:16.894 40cc Delete type=3 #1340
